requirements.txt
numpy
pillow
pandas
tensorflow
tqdm
scikit-learn
